feat(tangent): move explosion to std
