+++ guest/rust/Cargo.toml
@@ -69,6 +69,7 @@ members = [
     "packages/std/fps_controller",
     "packages/std/orbit_camera",
     "packages/std/hide_cursor",
+    "packages/std/explosion",
 
     # Packages (tools)
     "packages/tools/console",
