+++ guest/rust/packages/games/tangent/core/ambient.toml
@@ -7,8 +7,11 @@ content = { type = "Playable" }
 [dependencies]
 tangent_schema = { path = "../schema", deployment = "1MdfsKM0PneOuY19MlhaOw" }
 hide_cursor = { path = "../../../std/hide_cursor", deployment = "A8CV3CgLgnnNNSsr2m9kL" }
+explosion = { path = "../../../std/explosion" }
 
-kenney_impact_sounds = { path = "../../../assets/kenney_impact_sounds" , deployment = "5NieG6TwTyjNJMwgkdWLbZ" }
+game_object = { path = "../../../schemas/game_object" }
+
+kenney_impact_sounds = { path = "../../../assets/kenney_impact_sounds", deployment = "5NieG6TwTyjNJMwgkdWLbZ" }
 
 [messages.Input]
 name = "Input"
