+++ guest/rust/packages/games/tangent/mods/gun_laser/ambient.toml
@@ -6,6 +6,9 @@ content = { type = "Mod", for_playables = ["tangent"] }
 
 [dependencies]
 tangent_schema = { path = "../../schema", deployment = "1MdfsKM0PneOuY19MlhaOw" }
+
+explosion = { path = "../../../../std/explosion" }
+
 kenney_space_kit = { path = "../../../../assets/kenney_space_kit", deployment = "19EG2Sg0txG4WXz2lW73LL" }
 kenney_digital_audio = { path = "../../../../assets/kenney_digital_audio", deployment = "6fsl0CRjEUzYrHgXRh2YGK" }
