+++ guest/rust/packages/games/tangent/mods/pickup_health/ambient.toml
@@ -6,7 +6,8 @@ content = { type = "Mod", for_playables = ["tangent"] }
 
 [dependencies]
 tangent_schema = { path = "../../schema", deployment = "1MdfsKM0PneOuY19MlhaOw" }
-kenney_digital_audio = { path = "../../../../assets/kenney_digital_audio" , deployment = "6fsl0CRjEUzYrHgXRh2YGK" }
+kenney_digital_audio = { path = "../../../../assets/kenney_digital_audio", deployment = "6fsl0CRjEUzYrHgXRh2YGK" }
+game_object = { path = "../../../../schemas/game_object" }
 
 [components]
 is_health_pickup = { type = "Empty", name = "Is Health Pickup", description = "This entity is a health pickup." }
