+++ guest/rust/packages/games/tangent/mods/ui_holohud/ambient.toml
@@ -6,3 +6,4 @@ content = { type = "Mod", for_playables = ["tangent"] }
 
 [dependencies]
 tangent_schema = { path = "../../schema", deployment = "1MdfsKM0PneOuY19MlhaOw" }
+game_object = { path = "../../../../schemas/game_object" }
