+++ guest/rust/packages/std/explosion/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "explosion"
+
+edition = "2021"
+publish = false
+rust-version = {workspace = true}
+version = {workspace = true}
+
+[dependencies]
+ambient_api = { path = "../../../api" }
+
+[[bin]]
+name = "explosion_client"
+path = "src/client.rs"
+required-features = ["client"]
+
+[[bin]]
+name = "explosion_server"
+path = "src/server.rs"
+required-features = ["server"]
+
+[features]
+client = ["ambient_api/client"]
+server = ["ambient_api/server"]
